requirements.txt
tensorflow
opencv-python-headless
numpy
pillow
streamlit
scikit-learn
matplotlib
